feat: added Edit route tests
