+++ src/test/resources/scripts/controller/routes_before.sql
@@ -1,2 +1,1 @@
-INSERT INTO routes(id,name) VALUES
-    (1, 'Athlone');
+INSERT INTO route(id,name) VALUES (1, 'Athlone');
